+++ internal/server/handlers.go
@@ -2,7 +2,6 @@ package server
 
 import (
 	"encoding/json"
-	"errors"
 	"log"
 	"net/http"
 
@@ -20,13 +19,13 @@ func (s *Server) HandleCreatePlant(c *gin.Context) {
 		c.Writer.WriteHeader(http.StatusInternalServerError)
 		return
     }
+	if plant.Name == "" || plant.Description == "" {
+		log.Println("create request missing name or description")
+		c.Writer.WriteHeader(http.StatusBadRequest)
+		return
+	}
 	err = s.db.CreatePlant(plant, c)
 	if err != nil {
-		if errors.Is(err, err.(*db.ErrNotFound)) {
-			log.Println(err)
-			c.Writer.WriteHeader(http.StatusNotFound)
-			return
-		}
 		log.Println(err)
 		c.Writer.WriteHeader(http.StatusInternalServerError)
 		return
@@ -35,9 +34,14 @@ func (s *Server) HandleCreatePlant(c *gin.Context) {
 }
 
 func (s *Server) HandleGetPlant(c *gin.Context) {
+	if c.Param("name") == "" {
+		log.Println("create request missing name or description")
+		c.Writer.WriteHeader(http.StatusBadRequest)
+		return
+	}
 	plant, err := s.db.GetPlant(c.Param("name"), c)
 	if err != nil {
-		if errors.Is(err, db.ErrNotFound{}) {
+		if err.Error() == db.ErrNotFound {
 			log.Println(err)
 			c.Writer.WriteHeader(http.StatusNotFound)
 			return
@@ -58,6 +62,11 @@ func (s *Server) HandleUpdatePlant(c *gin.Context) {
 		c.Writer.WriteHeader(http.StatusInternalServerError)
 		return
     }
+	if plant.Name == "" || plant.Description == "" {
+		log.Println("create request missing name or description")
+		c.Writer.WriteHeader(http.StatusBadRequest)
+		return
+	}
 	err = s.db.UpdatePlant(plant, c)
 	if err != nil {
 		log.Println(err)
@@ -68,9 +77,14 @@ func (s *Server) HandleUpdatePlant(c *gin.Context) {
 }
 
 func (s *Server) HandleDeletePlant(c *gin.Context) {
+	if c.Param("name") == "" {
+		log.Println("create request missing name or description")
+		c.Writer.WriteHeader(http.StatusBadRequest)
+		return
+	}
 	plant, err := s.db.DeletePlant(c.Param("name"), c)
 	if err != nil {
-		if errors.Is(err, db.ErrNotFound{}) {
+		if err.Error() == db.ErrNotFound {
 			log.Println(err)
 			c.Writer.WriteHeader(http.StatusNotFound)
 			return
